==> Switch_task/01.generate_lt_df_Ali_8days.ipynb <==
import pandas as pd
import numpy as np
import datetime
import pickle
from tqdm import tqdm
import csv

dates = ['2017-11-26', '2017-11-27', '2017-11-28', '2017-11-29', '2017-11-30', '2017-12-01' , '2017-12-02', '2017-12-03'] # '2017-11-25', 

NN_tables = []
for part in range(1, 5):
    with open('../Taobao_v3/dict_NN_table/NN_table_'+ str(part) +'.pkl', 'rb') as file:
        NN_tables.append(pickle.load(file))

def NN_ranking(r):
    for part in range(4):
        if str(r.target_iid) in NN_tables[part]:
            lt = NN_tables[part][str(r.target_iid)]
            ids, score = zip(*lt)
            try:
                rank = ids.index(str(r.next_iid)) + 1
            except:
                rank = np.nan
            break
    else:
        rank = np.nan
    return rank

lt_df = []
for date in tqdm(dates):
    with open('../DIEN_T/testing_table/' + date + '_knn.pkl', 'rb') as file:
        df = pickle.load(file)
    df.dropna(inplace=True)
    df.reset_index(drop=True, inplace=True)
    df['NN_rank'] = df.apply(NN_ranking, axis=1)
    
    dff = pd.read_csv('./din_ali/' + 'results_' + date[5:] + '.csv')
    dff.rename(columns = {'rank':'DIN_rank'}, inplace=True)
    dff['uid'] = dff.apply(lambda df: eval(df['uid'])[0], axis=1)
    dff.sort_values(by=['uid'], inplace=True, ignore_index=True)
    dff.drop(columns=['hist_iid', 'last_iid', 'target_iid'], inplace=True)
    try:
        pd.testing.assert_series_equal(df['uid'], dff['uid'])
        df['DIN_rank'] = dff['DIN_rank']
    except Exception as e:
        print(date)
        print(e)
        df = pd.merge(df, dff, on=['uid'], how='left')
        df.fillna(dff['DIN_rank'].mean(), inplace=True)

    dff = pd.read_csv('./dien_ali/' + 'ali_dien_rank_' + date[5:] + '.csv')
    dff.rename(columns = {'dien_rank':'DIEN_rank'}, inplace=True)
#     dff['uid'] = dff.apply(lambda df: eval(df['uid'])[0], axis=1)
    dff.sort_values(by=['uid'], inplace=True, ignore_index=True)
    dff.drop(columns=['hist_iid', 'last_id', 'target_iid'], inplace=True)
    pd.testing.assert_series_equal(df['uid'], dff['uid'])
    df['DIEN_rank'] = dff['DIEN_rank']

    lt_df.append(df)

with open('./lt_df_Ali_8days.pkl', 'wb') as file:
    pickle.dump(lt_df, file)

lt_df[0]

lt_df[0].describe()
