==> Switch_task/02.generate_lt_df_dart.ipynb <==
import pandas as pd
import numpy as np
import datetime
import pickle
from tqdm import tqdm
import csv

from os import listdir
from os.path import isfile, isdir, join

# 指定要列出所有檔案的目錄
mypath = './dien_dart'
# 取得所有檔案與子目錄名稱
files = listdir(mypath)
csv_list = []
# 以迴圈處理
for f in files:
  # 產生檔案的絕對路徑
  fullpath = join(mypath, f)
  # 判斷 fullpath 是檔案還是目錄
  if isfile(fullpath):
#     print("檔案：", f)
    if f[-4:]=='.csv':
        csv_list.append(f[17:-4])
csv_list.sort()

csv_list

with open('../CTR_v4/item_data/dict_knn_table.pkl', 'rb') as file:
    dict_knn_table = pickle.load(file)

def knn_ranking(r):
    if r.target_iid in dict_knn_table:
        ids = dict_knn_table[r.target_iid]
#             ids, cos = zip(*lt)
        try:
            rank = ids.index(r.next_iid) + 1
        except:
            rank = np.nan
    else:
        rank = np.nan
    return rank

NN_tables = []
for part in range(1, 5):
    with open('../CTR_v4/dict_NN_table/NN_table_'+ str(part) +'.pkl', 'rb') as file:
        NN_tables.append(pickle.load(file))

def NN_ranking(r):
    for part in range(4):
        if r.target_iid in NN_tables[part]:
            lt = NN_tables[part][r.target_iid]
            ids, score = zip(*lt)
            try:
                rank = ids.index(r.next_iid) + 1
            except:
                rank = np.nan
            break
    else:
        rank = np.nan
    return rank

lt_df = []
for date in tqdm(csv_list):
    df = pd.read_csv('./dien_dart/' + 'change_dien_rank_' + date + '.csv').drop('Unnamed: 0', axis=1)    
    df.rename(columns = {'dien_rank':'DIEN_rank'}, inplace=True)
    df.sort_values(by=['uid'], inplace=True, ignore_index=True)
        
    dff = pd.read_csv('./din_dart/' + 'index_to_name_test_' + date + '_result' + '.csv')
    dff.rename(columns = {'din_rank':'DIN_rank'}, inplace=True)
    dff.sort_values(by=['uid'], inplace=True, ignore_index=True)
    dff.drop(columns=['hist_iid', 'last_iid', 'target_iid'], inplace=True)
    try:
        pd.testing.assert_series_equal(df['uid'], dff['uid'])
        df['DIN_rank'] = dff['DIN_rank']
    except Exception as e:
        print(date)
        print(e)
        df = pd.merge(df, dff, on=['uid'], how='left')
        df.fillna(dff['DIN_rank'].mean(), inplace=True)

    df.rename(columns = {'last_iid':'target_iid', 'target_iid':'next_iid'}, inplace=True)
    df['NN_rank'] = df.apply(NN_ranking, axis=1)
    df['knn_rank'] = df.apply(knn_ranking, axis=1)
    
    lt_df.append(df)

with open('./lt_df_dart_108days.pkl', 'wb') as file:
    pickle.dump(lt_df, file)

lt_df[0]

lt_df[100].describe()
