==> Switch_task/03.Ali_switch-F-XGB.ipynb <==
import pandas as pd
import numpy as np
import datetime
import time
import pickle
from tqdm import tqdm
import matplotlib.pyplot as plt
%matplotlib inline
from collections import Counter
from random import choices, sample
from gensim.models import Word2Vec
import gc
pd.set_option('display.float_format',lambda x : '%.2f' % x)
from math import log10

with open('./lt_df_Ali_8days.pkl', 'rb') as file:
    lt_df = pickle.load(file)

lt_df[0].columns

lt_df[0]

def gen_label(df):
    if df.knn_rank <= df.NN_rank and  df.knn_rank <= df.DIN_rank and df.knn_rank <= df.DIEN_rank:
        label = 0
    elif df.NN_rank <= df.knn_rank and  df.NN_rank <= df.DIN_rank and df.NN_rank <= df.DIEN_rank:
        label = 1
    elif df.DIN_rank <= df.knn_rank and  df.DIN_rank <= df.NN_rank and df.DIN_rank <= df.DIEN_rank:
        label = 2
    elif df.DIEN_rank <= df.knn_rank and  df.DIEN_rank <= df.NN_rank and df.DIEN_rank <= df.DIN_rank:
        label = 3
    return label

for df in tqdm(lt_df):
    df['label'] = df.apply(gen_label, axis=1)

for df in tqdm(lt_df):
    df['hist_len'] = df.apply(lambda df: len(eval(str(df['hist_iid']))), axis=1)

from keras.utils import to_categorical
for df in tqdm(lt_df):
    df['train_label'] = df.apply(lambda df: to_categorical(df['label'], num_classes=4), axis=1)

lt_df[0]

lt_df[0]['label'].value_counts()

lt_df[0][lt_df[0]['label']==0].describe()

lt_df[0][lt_df[0]['label']==1].describe()

lt_df[0][lt_df[0]['label']==2].describe()

lt_df[0][lt_df[0]['label']==3].describe()

---

full_table = pd.concat(lt_df[:4], ignore_index=True)

full_table['label'].value_counts()

full_table.describe()

table_grby = full_table.groupby(['target_iid'])
pd.DataFrame(table_grby.size())

df_table  = pd.merge(full_table, pd.DataFrame(table_grby.size()), on=['target_iid'], how='left')
df_table.rename(columns = {0:'occur'}, inplace=True)

max_occur = max(df_table['occur'])

df_table['occur_log'] = df_table.apply(lambda df: log10(df['occur'])/log10(max_occur), axis=1)

df_table

df_table.describe()

df_table.groupby('occur').mean()

---

model_w2v = Word2Vec.load("../Taobao_v3/iid_1days.model")

training_data = []
not_exist_id = []
Ground_Truth = []
for row in tqdm(df_table.iterrows()):
    if str(row[1]['target_iid']) not in model_w2v.wv:
        not_exist_id.append(row[1]['target_iid'])
        continue
    training_data.append(np.append(model_w2v.wv[str(row[1]['target_iid'])], [row[1]['occur_log']]))
    Ground_Truth.append(row[1]['label'])# train_label

training_data = np.array(training_data)

Ground_Truth = np.array(Ground_Truth)

training_data.shape

len(Ground_Truth)

from sklearn.model_selection import train_test_split
from keras import models, layers, optimizers, losses, metrics, callbacks

from imblearn.under_sampling import RandomUnderSampler, NearMiss, TomekLinks, NeighbourhoodCleaningRule
# X_train, y_train = training_data, Ground_Truth
# X_train, y_train = RandomUnderSampler().fit_resample(training_data, Ground_Truth)
# X_train, y_train = NearMiss(version=2).fit_resample(training_data, Ground_Truth)
X_train, y_train = TomekLinks(sampling_strategy='all', n_jobs=8).fit_resample(training_data, Ground_Truth)
# X_train, y_train = NeighbourhoodCleaningRule().fit_resample(training_data, Ground_Truth)

pd.DataFrame(Ground_Truth).sum()

pd.DataFrame(y_train).sum()

y_train[0]

from xgboost import XGBClassifier
model = XGBClassifier()
model.fit(X_train, y_train)

model.predict(np.append(model_w2v.wv['3'], 0.5).reshape(1, -1))[0]

model.save_model('./saved_model/Ali_switch_0714_F_XGB.model')

---

# model = models.load_model('./saved_model/Ali_switch_0524_F.h5')

df_table_testday = pd.concat(lt_df[4:], ignore_index=True)

# df_table_testday.drop_duplicates(inplace=True, ignore_index=True)

df_table_testday.describe()

df_table_testday = pd.merge(df_table_testday, pd.DataFrame(table_grby.size()), on=['target_iid'], how='left')

df_table_testday.rename(columns = {0:'occur'}, inplace=True)

df_table_testday['occur'].fillna(1, inplace=True)

# model.predict(np.append(model_w2v.wv['2063176'], 0.3).reshape(1, -1))[0]

# np.argmax(model.predict(np.append(model_w2v.wv['2063176'], 0.3).reshape(1, -1))[0])

# sum(model.predict(np.append(model_w2v.wv['2063176'], 0.3).reshape(1, -1))[0])

def pred(df):
    if str(df['target_iid']) not in model_w2v.wv:
        return 0
    else:
#         return np.argmax(model.predict(np.append(model_w2v.wv[str(df['target_iid'])], df['occur_log']).reshape(1, -1))[0])
        return model.predict(np.append(model_w2v.wv[str(df['target_iid'])], df['occur_log']).reshape(1, -1))[0]

def label_switch(df):
    label = df['pred']
    rank = None
    if label == 0:
        rank = df['knn_rank']
    elif label == 1:
        rank = df['NN_rank']
    elif label == 2:
        rank = df['DIN_rank']
    elif label == 3:
        rank = df['DIEN_rank']
    return rank

max_occur = max(df_table_testday['occur'])

df_table_testday['occur_log'] = df_table_testday.apply(lambda df: log10(df['occur'])/log10(max_occur), axis=1)

df_table_testday['pred'] = df_table_testday.apply(pred, axis=1)

df_table_testday['pred_rank'] = df_table_testday.apply(label_switch, axis=1)

df_table_testday.describe()

from sklearn.metrics import classification_report,confusion_matrix
print(classification_report(df_table_testday['label'].tolist(),df_table_testday['pred'].tolist()))

df_table_testday

y_train.tolist().count([0, 0, 0, 1])

---

len(df_table_testday[df_table_testday['knn_rank'] <= 20])

len(df_table_testday[df_table_testday['NN_rank'] <= 20])

len(df_table_testday[df_table_testday['DIN_rank'] <= 20])

len(df_table_testday[df_table_testday['DIEN_rank'] <= 20])

len(df_table_testday[df_table_testday['pred_rank'] <= 20])
